requirements.txt
crewai
groq
langchain-groq
python-dotenv
bcrypt
streamlit
streamlit-option-menu
